Create CT-Guided-Injections-For-Pain-Management-by-Crystal-Radiology-in-Auburn.-02-8315-8292
CT-guided injections are a modern procedure that uses imaging to precisely inject anti-inflammatory cortisone into different parts of the spine for pain relief. These injections are considered when patients still experience intolerable pain after oral medication and physiotherapy have been tried. They work by treating the inflammation of nerve roots or facet joints of the spine and any inflammation in joints throughout the body.
What Are CT Guided Injections?
A CT-guided injection may deliver cortisone at the site of pain. Cortisone belongs to a group of drugs commonly known as corticosteroids, which can treat pain in any part of the body caused by inflammation. When CT imaging is used to guide the injection into the area of pain, it allows for accurately delivering a high dose of cortisone without the potential side effects of a cortisone tablet.
How Is the Procedure Performed?
The procedure is relatively safe and is performed at most scan centres by radiologists or in operating theatres by surgeons and pain physicians. It is done under a local anesthetic, and at times mild sedation may be administered.
1.	Preparation: An antiseptic wash is used to clean the skin over the area, and a local anesthetic is provided using a fine needle. The patient may experience a slight sting for a short time before the area numbs.
2.	Injection: Using a CT for guidance, a fine needle will be inserted at the injection site. Patients may experience some mild discomfort due to the needle placement and during the injection of cortisone.
3.	Completion: Upon completion of the injection, the proceduralist will remove the needle and place an adhesive dressing at the injection site.
What Are the Risks Associated With This Procedure?
Adverse reactions or any serious complications are very rare. Life-threatening reactions are extremely rare, observed only in about 1 in 170,000 patients, making this injection considerably safe. The most common side effect is mild tenderness around the site of injection, which usually disappears within two days post-procedure.
Some patients may experience the following side effects:
•	Worsened pain symptoms and weakness for 1 to 2 days post-surgery
•	An increase in blood sugar levels for diabetic patients
•	Insomnia, palpitations, flushing
•	Soreness at the site of injection, local bruising
Rare but serious side effects include:
•	Infection
•	Permanent neurological damage
•	Allergic reactions to the x-ray contrast, such as flushing, difficulty in breathing, or hives
How to Prepare for a CT Guided Injection
Before the procedure:
•	Eat a light meal a few hours before the injection.
•	For insulin-dependent diabetic patients, doctors usually recommend not changing their normal eating patterns.
•	Review your medical history with your doctor, who may advise stopping certain medications before the injection.
•	Inform the radiographer if you are pregnant or suspect you may be pregnant.
After the procedure:
•	Do not drive a vehicle post-procedure. You can return to your normal activities the following day.
•	Pain relief might be immediate, but neck or back pain can recur once the anesthesia wears off. This is expected as the cortisone takes about 2 days to start showing its effects.
•	Avoid strenuous activities or exercise for about 2 weeks after the cortisone starts showing its effect.
•	Record the pain relief experienced post-procedure as advised by your doctor.
•	Follow up with your doctor two weeks post-procedure.
What Is the Success Rate of CT-guided injections?
CT-guided injections are mostly successful in providing pain relief to most patients. The effect of each injection tends to be temporary, ranging from one month to one year. Despite this, they are still considered beneficial in providing pain relief to patients, especially during a severe episode of back pain.
Why Choose Crystal Radiology?
At Crystal Radiology, we are a local, privately-owned medical clinic offering the highest quality medical and imaging services in the region. We are leaders in bulk billing, providing the real benefit of no 'Out of Pocket' expenses for you. Our commitment to patient care, combined with our advanced imaging technology, ensures that you receive the most accurate diagnosis and effective treatment plan.
Our team of experienced radiologists and medical professionals is dedicated to delivering personalized care tailored to your unique needs. We understand the importance of timely and precise pain management, and our CT-guided injections are designed to provide you with the relief you need to improve your quality of life.
Your Next Step
Please ensure that you have your Medicare card and referral with you and pop into our location for your consultation. If you have any questions at all, please feel free to contact us at (02) 8315 8292 as we are here to help. We look forward to assisting you on your journey to better health and well-being.